refactor: separate device->notifications to a separate module
