+++ src/api/devices/types.ts
@@ -1,21 +1,5 @@
 // SPDX-License-Identifier: ice License 1.0
 
-export type NotificationDeliveryChannel = 'email' | 'push';
-
-export type NotificationDomain = string;
-
-export type NotificationDomainToggle = {
-  type: NotificationDomain;
-  enabled: boolean;
-};
-
-export type NotificationDomainToggles = Array<NotificationDomainToggle>;
-export interface DeviceSettings extends DeviceId {
-  disableAllNotifications: boolean;
-  pushNotificationSettings: NotificationDomainToggles;
-  emailNotificationSettings: NotificationDomainToggles;
-}
-
 export interface DeviceLocation {
   city: string;
   country: string;
