+++ src/api/notifications/constants.ts
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: ice License 1.0
 
-import {NotificationDomain} from '@api/devices/types';
+import {NotificationDomain} from '@api/notifications/types';
 
 export const DISABLE_ALL_NOTIFICATION_DOMAIN: NotificationDomain =
   'disable_all';
