+++ src/store/modules/Devices/selectors/index.ts
@@ -1,7 +1,5 @@
 // SPDX-License-Identifier: ice License 1.0
 
-import {NotificationDeliveryChannel} from '@api/devices/types';
-import {DISABLE_ALL_NOTIFICATION_DOMAIN} from '@api/notifications/constants';
 import {RootState} from '@store/rootReducer';
 
 export const deviceUniqueIdSelector = (state: RootState) =>
@@ -12,34 +10,3 @@ export const deviceLocationSelector = (state: RootState) =>
 
 export const lastMetadataUpdateSelector = (state: RootState) =>
   state.devices.lastMetadataUpdateAt;
-
-export const emailNotificationByTypeSelector = (state: RootState) => {
-  return state.devices.settings?.emailNotificationSettings;
-};
-
-export const pushNotificationByTypeSelector = (state: RootState) => {
-  return state.devices.settings?.pushNotificationSettings;
-};
-
-export const enabledNotificationDomainsSelector =
-  (channel: NotificationDeliveryChannel) => (state: RootState) => {
-    const notificationDomainToggles =
-      channel === 'push'
-        ? state.devices.settings?.pushNotificationSettings
-        : state.devices.settings?.emailNotificationSettings;
-    if (notificationDomainToggles) {
-      const enabledDomains = [];
-      for (const {type, enabled} of notificationDomainToggles) {
-        if (type === DISABLE_ALL_NOTIFICATION_DOMAIN) {
-          if (enabled) {
-            return [];
-          }
-        } else if (enabled) {
-          enabledDomains.push(type);
-        }
-      }
-
-      return enabledDomains;
-    }
-    return [];
-  };
