+++ src/store/modules/Notifications/sagas/updateNotificationChannel.ts
@@ -2,13 +2,13 @@
 
 import {Api} from '@api/index';
 import {AnalyticsEventLogger} from '@store/modules/Analytics/constants';
-import {DeviceActions} from '@store/modules/Devices/actions';
+import {NotificationActions} from '@store/modules/Notifications/actions';
 import {getErrorMessage} from '@utils/errors';
 import {call, put} from 'redux-saga/effects';
 
 export function* updateNotificationChannel(
   action: ReturnType<
-    typeof DeviceActions.UPDATE_NOTIFICATION_CHANNEL.START.create
+    typeof NotificationActions.UPDATE_NOTIFICATION_CHANNEL.START.create
   >,
 ) {
   const {notificationChannel, notificationDeliveryChannel} = action.payload;
@@ -20,7 +20,7 @@ export function* updateNotificationChannel(
     );
 
     yield put(
-      DeviceActions.UPDATE_NOTIFICATION_CHANNEL.SUCCESS.create(
+      NotificationActions.UPDATE_NOTIFICATION_CHANNEL.SUCCESS.create(
         notificationChannel,
         notificationDeliveryChannel,
       ),
@@ -31,7 +31,7 @@ export function* updateNotificationChannel(
     });
   } catch (error) {
     yield put(
-      DeviceActions.UPDATE_NOTIFICATION_CHANNEL.FAILED.create(
+      NotificationActions.UPDATE_NOTIFICATION_CHANNEL.FAILED.create(
         getErrorMessage(error),
       ),
     );
